% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{fit}
\alias{fit}
\title{Predict Generic Function
This declares a generic function \code{predict}}
\usage{
fit(obj)
}
\description{
Predict Generic Function
This declares a generic function \code{predict}
}
