% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{nearest_classifier}
\alias{nearest_classifier}
\title{Simple Nearest Neighbors Classifier
It returns a classifier configured to classify}
\usage{
nearest_classifier()
}
\arguments{
\item{x}{Data points(Features of the data)}

\item{y}{Labels of the data}
}
\description{
Simple Nearest Neighbors Classifier
It returns a classifier configured to classify
}
\keyword{nearest}
\keyword{neighbors}
