% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbors.R
\name{predict.knn}
\alias{predict.knn}
\title{Predicts a data point}
\usage{
\method{predict}{knn}(knn, x)
}
\arguments{
\item{x}{The data points to predict}

\item{obj}{The model}
}
\description{
Predicts a data point
}
