% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear.R
\name{print.linear_regressor}
\alias{print.linear_regressor}
\title{Generic Print Function for Linear Regressor}
\usage{
\method{print}{linear_regressor}(x)
}
\description{
Generic Print Function for Linear Regressor
}
