% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{score}
\alias{score}
\title{Predict Generic Function
This declares a generic function \code{predict}
Score Generic Function
This declares a generic function \code{score}}
\usage{
score(obj, ...)
}
\description{
Predict Generic Function
This declares a generic function \code{predict}
Score Generic Function
This declares a generic function \code{score}
}
